textminingapi/requirements.txt
django
openai
djangorestframework
python-dotenv
